organized methods into areas of the record
