+++ lib/marc_cleanup.rb
@@ -2,11 +2,10 @@ require_relative 'marc_cleanup/constants'
 require_relative 'marc_cleanup/unicode_repertoire'
 require_relative 'marc_cleanup/rawmarc_errors'
 require_relative 'marc_cleanup/fixed_fields'
-require_relative 'marc_cleanup/global_errors'
-require_relative 'marc_cleanup/oclc_validation'
+require_relative 'marc_cleanup/record_level'
 require_relative 'marc_cleanup/variable_fields'
 require_relative 'marc_cleanup/bib_marc_fixes'
-require_relative 'marc_cleanup/bib_rubymarc_fixes'
+require_relative 'marc_cleanup/leader'
 require 'marc'
 require 'uri'
 include MarcCleanup
