+++ spec/fixtures/record_level/marc_with_duplicate_fields.xml
