+++ spec/record_level/repeatable_field_errors_spec.rb
@@ -5,7 +5,7 @@ require 'marc_cleanup'
 
 RSpec.describe 'repeatable_field_errors?' do
   describe 'repeated non-repeatable field' do
-    let(:marcfile) { "#{Dir.getwd}/spec/fixtures/global_errors/marc_with_multiple_001.xml" }
+    let(:marcfile) { "#{Dir.getwd}/spec/fixtures/record_level/marc_with_multiple_001.xml" }
     let(:reader) { MARC::XMLReader.new(marcfile, parser: 'magic') }
     let(:record_with_repeatable_field_errors) { reader.first }
     it 'finds repeatable field errors in the record' do
